requirements.txt
python-dotenv == 0.20.0
matplotlib == 3.5.2
discord.py == 1.7.3
requests == 2.27.1
